+++ CHANGELOG.md
@@ -1,5 +1,17 @@
 # Changelog
 
+## Unreleased
+
+#### 🚀 Updates
+
+- Added support for `#[serde(deny_unknown_fields)]` to struct and enum containers.
+- Added support for `#[serde(rename_all_fields)]` to enum containers.
+- Added support for `#[serde(untagged)]` to enum variants.
+
+#### ⚙️ Internal
+
+- Updated dependencies.
+
 ## 0.17.3
 
 #### 🐞 Fixes